Fixed waste card dragging issue
When 2 or more cards were stacked in the waste pile, picking the
top one up caused the card underneath to change containers from
a simple div to a DropTarget. This caused an issue with Reconciliation[1]
where the DOM update caused the Draggable which had the in-progress
dragging state to be unmounted, and a new Draggable mounted in its place,
losing the state.

This change ensures that the underlying waste card does not get a
DropTarget container, which is accurate to the rules regardless of
the bug. However, I need to be on the look out for similar issues.

It was helpful to log componentDidMount/componentDidUnmount to watch
the lifecycle boundaries of the Draggable.

[1]: http://facebook.github.io/react/docs/reconciliation.html