Don't swap containers between Draggable and DropTarget
This cause the DOM subtrees to be discarded and reloaded, which
resulted in a flicker as the <img> for the faces redownloaded.

Reloading the subtree is an artifact of how Reconciliation works
in React. This is my second run-in with Draggable and DropTarget
and Reconciliation (ff7e6fdc4807ed8ee2d552098358b2ff036e012f).

The solution here is to merge the two classes together, to create
`DragAndDrop`. It inspects the properties to determine which
behaviors are active. The Board top-level function `containerForCard`
has been replaced with `isCardDraggable` and `isCardDropTarget`, which
are called by Stack. Every stack is now a `DragAndDrop`, where as
before some cards would be plain `div`s.